services/real-estate-services/requirements.txt
numpy
pandas
requests
flask
gunicorn
scikit-learn==1.6.1
joblib
google-cloud-storage